refactor: ApartmentXmlMapper 클래스 위치 이동
